Make role/aria-* as HTML attribute, not DOM prop
Firefox, as of writing this commit message, does not support either Element.role or
Element.aria* properties. So, values set in rescript-tea are not
reflected in Firefox.

For example, if role is set as `Vdom.prop()`, then Firefox renders

```
ol(list{role("tablist")}, list{})
```

simply as

```
<ol></ol>
```

On the other hand, if role is set as `Vdom.attribute()`, then Firefox
correctly renders it as

```
<ol role="tablist"></ol>
```